HBASE-22874 Define a public API for Canary checking and a non-public tool implementation
Closes #580

* Canary is now an IA.Public interface
* CanaryTool is now the implementation

(cherry picked from commit b5a396749d150b0aa3cd8a5eadaaf124a0859c39)
