+++ bin/hbase
@@ -607,7 +607,7 @@ elif [ "$COMMAND" = "ltt" ] ; then
   CLASS='org.apache.hadoop.hbase.util.LoadTestTool'
   HBASE_OPTS="$HBASE_OPTS $HBASE_LTT_OPTS"
 elif [ "$COMMAND" = "canary" ] ; then
-  CLASS='org.apache.hadoop.hbase.tool.Canary'
+  CLASS='org.apache.hadoop.hbase.tool.CanaryTool'
   HBASE_OPTS="$HBASE_OPTS $HBASE_CANARY_OPTS"
 elif [ "$COMMAND" = "version" ] ; then
   CLASS='org.apache.hadoop.hbase.util.VersionInfo'
